starting API stuff
